--- a/Base32768.Benchmark/Base32768.Benchmark.csproj
+++ b/Base32768.Benchmark/Base32768.Benchmark.csproj
@@ -2,8 +2,8 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <LangVersion>9</LangVersion>
-    <TargetFrameworks>net5.0;net472</TargetFrameworks>
+    <LangVersion>10</LangVersion>
+    <TargetFrameworks>net6.0;net5.0;net472</TargetFrameworks>
     <RootNamespace>Kzrnm.Convert.Base32768.Benchmark</RootNamespace>
 
     <NoWarn>CA1050</NoWarn>