--- a/Base32768/Base32768.csproj
+++ b/Base32768/Base32768.csproj
@@ -1,6 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 	<PropertyGroup>
-		<TargetFrameworks>net35;netstandard1.0;netstandard2.0;netstandard2.1</TargetFrameworks>
+		<TargetFrameworks>net35;net45;netstandard1.0;netstandard2.0;netstandard2.1</TargetFrameworks>
 		<LangVersion>latest</LangVersion>
 		<GeneratePackageOnBuild Condition="'$(Configuration)' == 'Release'">true</GeneratePackageOnBuild>