--- a/Base32768/Utils.cs
+++ b/Base32768/Utils.cs
@@ -1,7 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Runtime.CompilerServices;
-using System.Text;
 
 #if !NETCOREAPP3_0_OR_GREATER
 namespace System.Runtime.CompilerServices