Fix #11 

``` ini

BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19043.1348 (21H1/May2021Update)
Intel Core i7-4790 CPU 3.60GHz (Haswell), 1 CPU, 8 logical and 4 physical cores
.NET SDK=6.0.100
  [Host]   : .NET 6.0.0 (6.0.21.52210), X64 RyuJIT
  ShortRun : .NET 5.0.12 (5.0.1221.52207), X64 RyuJIT

Job=ShortRun  IterationCount=3  LaunchCount=1  
WarmupCount=3  

```
|    Method |            Toolchain |      N |       Mean |       Error |     StdDev |   Gen 0 |   Gen 1 |   Gen 2 | Allocated |
|---------- |--------------------- |------- |-----------:|------------:|-----------:|--------:|--------:|--------:|----------:|
| **OldEncode** |             **.NET 5.0** |   **1000** |   **2.604 μs** |   **1.0034 μs** |  **0.0550 μs** |  **0.5569** |       **-** |       **-** |      **2 KB** |
| OldEncode |             .NET 6.0 |   1000 |   1.880 μs |   0.1959 μs |  0.0107 μs |  0.5569 |       - |       - |      2 KB |
| OldEncode | .NET Framework 4.7.2 |   1000 |   2.527 μs |   1.0244 μs |  0.0562 μs |  0.5608 |       - |       - |      2 KB |
| **OldEncode** |             **.NET 5.0** | **100000** | **245.114 μs** |  **35.1314 μs** |  **1.9257 μs** | **66.4063** | **66.4063** | **66.4063** |    **209 KB** |
| OldEncode |             .NET 6.0 | 100000 | 202.086 μs | 103.3177 μs |  5.6632 μs | 66.6504 | 66.6504 | 66.6504 |    209 KB |
| OldEncode | .NET Framework 4.7.2 | 100000 | 256.566 μs | 292.9717 μs | 16.0588 μs | 66.6504 | 66.6504 | 66.6504 |    209 KB |
|           |                      |        |            |             |            |         |         |         |           |
| **NewEncode** |             **.NET 5.0** |   **1000** |   **2.590 μs** |   **0.2851 μs** |  **0.0156 μs** |  **0.5455** |       **-** |       **-** |      **2 KB** |
| NewEncode |             .NET 6.0 |   1000 |   1.861 μs |   0.3724 μs |  0.0204 μs |  0.5455 |       - |       - |      2 KB |
| NewEncode | .NET Framework 4.7.2 |   1000 |   2.337 μs |   0.3851 μs |  0.0211 μs |  0.5531 |       - |       - |      2 KB |
| **NewEncode** |             **.NET 5.0** | **100000** | **247.493 μs** |  **19.4350 μs** |  **1.0653 μs** | **66.4063** | **66.4063** | **66.4063** |    **208 KB** |
| NewEncode |             .NET 6.0 | 100000 | 189.162 μs |  67.3899 μs |  3.6939 μs | 66.6504 | 66.6504 | 66.6504 |    209 KB |
| NewEncode | .NET Framework 4.7.2 | 100000 | 238.618 μs | 299.9674 μs | 16.4422 μs | 66.6504 | 66.6504 | 66.6504 |    209 KB |
|           |                      |        |            |             |            |         |         |         |           |
| **OldDecode** |             **.NET 5.0** |   **1000** |   **1.659 μs** |   **0.4437 μs** |  **0.0243 μs** |  **0.2441** |       **-** |       **-** |      **1 KB** |
| OldDecode |             .NET 6.0 |   1000 |   1.663 μs |   0.2782 μs |  0.0152 μs |  0.2441 |       - |       - |      1 KB |
| OldDecode | .NET Framework 4.7.2 |   1000 |   1.741 μs |   0.5883 μs |  0.0322 μs |  0.2441 |       - |       - |      1 KB |
| **OldDecode** |             **.NET 5.0** | **100000** | **174.114 μs** |  **38.5285 μs** |  **2.1119 μs** | **31.0059** | **31.0059** | **31.0059** |     **98 KB** |
| OldDecode |             .NET 6.0 | 100000 | 179.637 μs |  44.2308 μs |  2.4244 μs | 31.0059 | 31.0059 | 31.0059 |     98 KB |
| OldDecode | .NET Framework 4.7.2 | 100000 | 194.217 μs | 174.9644 μs |  9.5904 μs | 31.0059 | 31.0059 | 31.0059 |     98 KB |
|           |                      |        |            |             |            |         |         |         |           |
| **NewDecode** |             **.NET 5.0** |   **1000** |   **1.572 μs** |   **0.2770 μs** |  **0.0152 μs** |  **0.2441** |       **-** |       **-** |      **1 KB** |
| NewDecode |             .NET 6.0 |   1000 |   1.918 μs |   4.0247 μs |  0.2206 μs |  0.2441 |       - |       - |      1 KB |
| NewDecode | .NET Framework 4.7.2 |   1000 |   1.871 μs |   4.1859 μs |  0.2294 μs |  0.2441 |       - |       - |      1 KB |
| **NewDecode** |             **.NET 5.0** | **100000** | **168.900 μs** | **111.2158 μs** |  **6.0961 μs** | **31.0059** | **31.0059** | **31.0059** |     **98 KB** |
| NewDecode |             .NET 6.0 | 100000 | 160.536 μs |  58.8320 μs |  3.2248 μs | 31.0059 | 31.0059 | 31.0059 |     98 KB |
| NewDecode | .NET Framework 4.7.2 | 100000 | 201.303 μs | 169.0498 μs |  9.2662 μs | 31.0059 | 31.0059 | 31.0059 |     98 KB |