@firepacket 2021-11-13
Holy crap you're encoding 100k at 250 nano seconds and decoding at 250 
nanoseconds? You just did that today?? Wow where did you learn that 
trick? You must be younger than me, I resist all these new frameworks 
but you're showing me it really pays to start learning.

Make the numbers bigger. Do

On 11/13/2021 12:52 AM, kzrnm wrote:

> Fix #11 [1]
> 
> BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19043.1348 
> (21H1/May2021Update)
> Intel Core i7-4790 CPU 3.60GHz (Haswell), 1 CPU, 8 logical and 4 
> physical cores
> .NET SDK=6.0.100
> [Host]   : .NET 6.0.0 (6.0.21.52210), X64 RyuJIT
> ShortRun : .NET 5.0.12 (5.0.1221.52207), X64 RyuJIT
> 
> Job=ShortRun  IterationCount=3  LaunchCount=1
> WarmupCount=3
> 
> METHOD
> TOOLCHAIN
> N
> MEAN
> ERROR
> STDDEV
> GEN 0
> GEN 1
> GEN 2
> ALLOCATED
> 
> OldEncode
> .NET 5.0
> 1000
> 2.604 μs
> 1.0034 μs
> 0.0550 μs
> 0.5569
> -
> -
> 2 KB
> 
> OldEncode
> .NET 6.0
> 1000
> 1.880 μs
> 0.1959 μs
> 0.0107 μs
> 0.5569
> -
> -
> 2 KB
> 
> OldEncode
> .NET Framework 4.7.2
> 1000
> 2.527 μs
> 1.0244 μs
> 0.0562 μs
> 0.5608
> -
> -
> 2 KB
> 
> OldEncode
> .NET 5.0
> 100000
> 245.114 μs
> 35.1314 μs
> 1.9257 μs
> 66.4063
> 66.4063
> 66.4063
> 209 KB
> 
> OldEncode
> .NET 6.0
> 100000
> 202.086 μs
> 103.3177 μs
> 5.6632 μs
> 66.6504
> 66.6504
> 66.6504
> 209 KB
> 
> OldEncode
> .NET Framework 4.7.2
> 100000
> 256.566 μs
> 292.9717 μs
> 16.0588 μs
> 66.6504
> 66.6504
> 66.6504
> 209 KB
> 
> NewEncode
> .NET 5.0
> 1000
> 2.590 μs
> 0.2851 μs
> 0.0156 μs
> 0.5455
> -
> -
> 2 KB
> 
> NewEncode
> .NET 6.0
> 1000
> 1.861 μs
> 0.3724 μs
> 0.0204 μs
> 0.5455
> -
> -
> 2 KB
> 
> NewEncode
> .NET Framework 4.7.2
> 1000
> 2.337 μs
> 0.3851 μs
> 0.0211 μs
> 0.5531
> -
> -
> 2 KB
> 
> NewEncode
> .NET 5.0
> 100000
> 247.493 μs
> 19.4350 μs
> 1.0653 μs
> 66.4063
> 66.4063
> 66.4063
> 208 KB
> 
> NewEncode
> .NET 6.0
> 100000
> 189.162 μs
> 67.3899 μs
> 3.6939 μs
> 66.6504
> 66.6504
> 66.6504
> 209 KB
> 
> NewEncode
> .NET Framework 4.7.2
> 100000
> 238.618 μs
> 299.9674 μs
> 16.4422 μs
> 66.6504
> 66.6504
> 66.6504
> 209 KB
> 
> OldDecode
> .NET 5.0
> 1000
> 1.659 μs
> 0.4437 μs
> 0.0243 μs
> 0.2441
> -
> -
> 1 KB
> 
> OldDecode
> .NET 6.0
> 1000
> 1.663 μs
> 0.2782 μs
> 0.0152 μs
> 0.2441
> -
> -
> 1 KB
> 
> OldDecode
> .NET Framework 4.7.2
> 1000
> 1.741 μs
> 0.5883 μs
> 0.0322 μs
> 0.2441
> -
> -
> 1 KB
> 
> OldDecode
> .NET 5.0
> 100000
> 174.114 μs
> 38.5285 μs
> 2.1119 μs
> 31.0059
> 31.0059
> 31.0059
> 98 KB
> 
> OldDecode
> .NET 6.0
> 100000
> 179.637 μs
> 44.2308 μs
> 2.4244 μs
> 31.0059
> 31.0059
> 31.0059
> 98 KB
> 
> OldDecode
> .NET Framework 4.7.2
> 100000
> 194.217 μs
> 174.9644 μs
> 9.5904 μs
> 31.0059
> 31.0059
> 31.0059
> 98 KB
> 
> NewDecode
> .NET 5.0
> 1000
> 1.572 μs
> 0.2770 μs
> 0.0152 μs
> 0.2441
> -
> -
> 1 KB
> 
> NewDecode
> .NET 6.0
> 1000
> 1.918 μs
> 4.0247 μs
> 0.2206 μs
> 0.2441
> -
> -
> 1 KB
> 
> NewDecode
> .NET Framework 4.7.2
> 1000
> 1.871 μs
> 4.1859 μs
> 0.2294 μs
> 0.2441
> -
> -
> 1 KB
> 
> NewDecode
> .NET 5.0
> 100000
> 168.900 μs
> 111.2158 μs
> 6.0961 μs
> 31.0059
> 31.0059
> 31.0059
> 98 KB
> 
> NewDecode
> .NET 6.0
> 100000
> 160.536 μs
> 58.8320 μs
> 3.2248 μs
> 31.0059
> 31.0059
> 31.0059
> 98 KB
> 
> NewDecode
> .NET Framework 4.7.2
> 100000
> 201.303 μs
> 169.0498 μs
> 9.2662 μs
> 31.0059
> 31.0059
> 31.0059
> 98 KB
> 
> -------------------------
> 
> YOU CAN VIEW, COMMENT ON, OR MERGE THIS PULL REQUEST ONLINE AT:
> 
> https://github.com/kzrnm/Base32768/pull/12
> 
> COMMIT SUMMARY
> 
> * ff19173 [2] Add AggressiveInlining
> * 24462f7 [3] Support .NET Framework 4.5
> * 41f1819 [4] Remove unused usings
> * 802be4c [5] Benchmark old and new
> * d3a033f [6] Revert "Benchmark old and new"
> 
> FILE CHANGES
> 
> (5 files [7])
> 
> * M Base32768/Base32768.Decode.cs [8] (5)
> * M Base32768/Base32768.Encode.cs [9] (2)
> * M Base32768/Base32768.cs [10] (6)
> * M Base32768/Base32768.csproj [11] (2)
> * M Base32768/Utils.cs [12] (2)
> 
> PATCH LINKS:
> 
> * https://github.com/kzrnm/Base32768/pull/12.patch
> * https://github.com/kzrnm/Base32768/pull/12.diff
> 
> --
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub [13], or unsubscribe 
> [14].
> Triage notifications on the go with GitHub Mobile for iOS [15] or 
> Android [16].


Links:
------
[1] https://github.com/kzrnm/Base32768/issues/11
[2] 
https://github.com/kzrnm/Base32768/pull/12/commits/ff19173556a01389854f7243af6b9585d57d9422
[3] 
https://github.com/kzrnm/Base32768/pull/12/commits/24462f72598e52a382a6c5d3d52ba0098b69e37f
[4] 
https://github.com/kzrnm/Base32768/pull/12/commits/41f1819f41035d680dc65aa6a78a792a6e896012
[5] 
https://github.com/kzrnm/Base32768/pull/12/commits/802be4cb4e4b39d89e7d27858cf57101b3384c72
[6] 
https://github.com/kzrnm/Base32768/pull/12/commits/d3a033fd5c03ae0e10f568019c5d626e0683d517
[7] https://github.com/kzrnm/Base32768/pull/12/files
[8] 
https://github.com/kzrnm/Base32768/pull/12/files#diff-316f299c898010eee2b98d5c89e31439dc62fb066e60a1685c6202e7f2dc8f67
[9] 
https://github.com/kzrnm/Base32768/pull/12/files#diff-f3103f9c54fe05c110414570c4a4bd061d1862fcbc523c16ecbde5969c5c4531
[10] 
https://github.com/kzrnm/Base32768/pull/12/files#diff-c0d6b131c5014a5cae986b0e2fe78ec52936815ab5c663e3fc4f2b25195368f0
[11] 
https://github.com/kzrnm/Base32768/pull/12/files#diff-abdddad6a94a29a667254735c034fd51f0a9814a501ce50935b1f7f2723a3b1b
[12] 
https://github.com/kzrnm/Base32768/pull/12/files#diff-d9558729221bac5a571b86dd9f19bd3b34c0aec767871f136ba05b3bf4b0e43f
[13] https://github.com/kzrnm/Base32768/pull/12
[14] 
https://github.com/notifications/unsubscribe-auth/AA5OOMVMSD6F3BQU5VVUICDULX4K3ANCNFSM5H6H6U5Q
[15] 
https://apps.apple.com/app/apple-store/id1477376905?ct=notification-email&amp;mt=8&amp;pt=524675
[16] 
https://play.google.com/store/apps/details?id=com.github.android&amp;referrer=utm_campaign%3Dnotification-email%26utm_medium%3Demail%26utm_source%3Dgithub